+++ src/ui/vanilla/entry-point/index.ts
@@ -1,6 +1,3 @@
-// before tauri v2.5 this script was done as a workaround to https://github.com/tauri-apps/tauri/issues/12348
-// but was fixed on https://github.com/tauri-apps/wry/pull/1531 so now this script is used to initialize the widgets system
-
 import "../../../../libs/core/styles/colors.css";
 import "../../../../libs/core/styles/spacings.css";
 import "../../../../libs/core/styles/shadows.css";